// src/utils/LE/globals/enums.ts
export enum EnEstadisTipoRegistro {
  cambioParametros = 1,
  estadisticos = 2,
  alarmas = 3,
  eventos = 4,
  debug = 5,
}

// --------------------------------------- EnTipoDato ---------------------------------------
export enum EnTipoDato {
  noVariable = 0,
  uint8 = 1,
  int8 = 2,
  uint16 = 3,
  int16 = 4,
  uint32 = 5,
  int32 = 6,
  float = 7,
  tiempo = 8,

  pUint8 = 9,
  pInt8 = 10,
  pUint16 = 11,
  pInt16 = 12,
  pUint32 = 13,
  pInt32 = 14,

  pFloat = 15,
  pTiempo = 16,

  fecha = 26,

  string4 = 30,
  string32 = 34,

  // ? Este tipo de dato está en desuso: se utiliza TIPO_P_STRING_unicode16_simple que está más optimizado
  pString32 = 36,

  pStringUnicode16 = 44,
  stringUnicode16 = 45,

  concatenado = 46,
  pStringUnicode32 = 47,

  stringUnicode32 = 48,

  int64 = 202,
  pUint64 = 205,
  pFecha = 206,
}

// --------------------------------------- EnTipoTrama ---------------------------------------
export enum EnTipoTrama {
  // TT
  omegaPantallaPlaca = 13, // TT_OMEGA_PANTALLA_PLACA (Pintado de pantallas)
  depuracion = 14, // TT_DEPURACION (Tramas depuración)
  serviciosClaveValor = 23, // TT_SERVICIOS_CLAVE_VALOR
  actualizacionServer = 24, // TT_ACTUALIZACION_SERVER (Actualización remota)
  sistema = 25, // TT_SISTEMA (Funciones de sistema)
  estadisticos = 26, // TT_ESTADISTICOS (Envío de estadísticos)
  comuniRadar = 27, // TT_COMUNI_RADAR (Comunicación radar)
  comuniBle = 28, // TT_COMUNI_BLE (Comunicación BLE)
  descargaFicherosFlash = 29, // TT_DESCARGA_FICHEROS_FLASH
  importacionExportacion = 30, // TT_IMPORTACION_EXPORTACION
  descargaSubidaFicheros = 31, // TT_DESCARGA_SUBIDA_FICHEROS
  actualizacionV2 = 32, // TT_ACTUALIZACION_V2
}
// ...
export enum EnTmEstadisticos {
  enviaEstadistico = 1, // TM_ESTADISTICOS_envia_estadistico
  rtEstadistico = 2, // TM_ESTADISTICOS_rt_estadistico
}
// ...
export enum EnAlarmasAccion {
  off = 0,
  onAlarma = 1, // activa relé asociado
  onAviso = 2, // no activa relé asociado
}

/** Estado instantáneo de la alarma (ON/OFF) */
export enum EnAlarmaEstado {
  off = 0,
  on = 1,
}
// ...
export enum EnEeEventosApli {
  inicioCrianza = 1,
  entradaAnimales = 2,
  finCrianza = 3,
  altaBajaRetirada = 4,
}